+++ CHANGES.md
@@ -4,6 +4,7 @@
 * add guard to `set_readme_info()`
 * add function to check timeout validity
 * re-use valid cache timeout
+* refactor getting cache key to `get_cache_key()`
 
 #### 12.23.1 / 2026-02-12
 * add guard to release asset development download link in REST API